+++ src/cmd/jec/GlassCatInstall.ts
@@ -1,0 +1,79 @@
+//  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+//
+//   Copyright 2016-2017 Pascal ECHEMANN.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+import {Command} from "../Command";
+import {TarballUtil} from "../../utils/TarballUtil";
+import * as fs from "fs-extra";
+import * as path from "path";
+import * as childProcess from "child_process";
+
+/**
+ * The <code>GlassCatInstall</code> command allows to install a new GlassCat
+ * server instance.
+ */
+export class GlassCatInstall implements Command {
+  
+  //////////////////////////////////////////////////////////////////////////////
+  // Constructor function
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Creates a new <code>GlassCatInstall</code> instance.
+   */
+  constructor() {}
+
+  //////////////////////////////////////////////////////////////////////////////
+  // Public methods
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * @inheritDoc
+   */
+  public run(argv:any):void {
+    let currentPath:string = process.cwd();
+    let util:TarballUtil = new TarballUtil();
+    let execProcess:childProcess.ChildProcess = null;
+    let tmpFolder:string = "jec-glasscat";
+    util.download(
+      "https://registry.npmjs.org/jec-glasscat/-/jec-glasscat-0.0.7.tgz",
+      currentPath,
+      (e:any)=> {
+        if(e === null) {
+          fs.copy(path.join(currentPath, tmpFolder), currentPath, (err:Error) => {
+            if(err) return console.error(err);
+            console.log('success!');
+            fs.remove(tmpFolder).then(() => {
+              console.log('temp folder removed');
+            }).catch((reason:any) => {
+              console.log('temp folder not removed:' + reason);
+            })
+            execProcess = childProcess.exec("npm install");
+            execProcess.stdout.on("data", function(data) {
+              console.log(data); 
+            });
+            execProcess.stderr.on("data", function(data) {
+              console.log(data); 
+            });
+            execProcess.on("exit", function(code, signal) {
+              console.log('child process exited with ' +
+              `code ${code} and signal ${signal}`); 
+            });
+          });
+        }
+      }
+    );
+  }
+}